ai_cource_builder/requirements.txt
streamlit
PyPDF2
ollama
google-api-python-client
python-dotenv